requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Mailman==0.3.0
email-validator==2.1.0
pillow==10.0.0
werkzeug==2.3.7
wtforms==3.1.1
python-dotenv==1.0.0
itsdangerous==2.1.2
Flask-Migrate==4.0.7
gunicorn==21.2.0
pytest==7.4.0
pytest-flask==1.2.0
sqlalchemy==2.0.27 
Flask-Migrate   
firebase_admin
requests==2.31.0
ntplib==0.4.0
